+++ getl/blocks/load/entrypoint.py
@@ -126,7 +126,9 @@ def batch_xml(conf: BlockConfig) -> DataFrame:
     :param str Alias=: an alias for the dataframe that is loaded
     :param str Suffix=.xml: the suffix of the file
     :param int BatchSize=200: the amount of data to process in one go
-    :param str JsonSchemaPath=: the file schema in json format
+    :param str JsonSchemaPath=: path to a file with schema in json format, if no path is submitted, inferSchema will be set to true
+    :param dict JsonSchema=: the file schema in json format, if no schema is submitted, inferSchema will be set to true
+    :param StructType PySparkSchema=: the file schema in PySpark StructType format, if no schema is submitted, inferSchema will be set to true
 
     ```
     SectionName:
@@ -139,6 +141,8 @@ def batch_xml(conf: BlockConfig) -> DataFrame:
             Suffix: .xml
             BatchSize: 200
             JsonSchemaPath: s3://bucket-name/schema.json
+            JsonSchema: dict
+            PySparkSchema: StructType
     ```
 
     """
@@ -163,9 +167,15 @@ def batch_xml(conf: BlockConfig) -> DataFrame:
     options = {"rowTag": conf.get("RowTag")}
     # Get the file schema
     schema_path = conf.get("JsonSchemaPath", None)
+    schema = conf.get("JsonSchema", None)
+    pyspark_schema = conf.get("PySparkSchema", None)
     if schema_path:
         schema = json.loads(S3Path(schema_path).read_text())
         options["schema"] = json_to_spark_schema(schema)
+    elif schema:
+        options["schema"] = json_to_spark_schema(schema)
+    elif pyspark_schema:
+        options["schema"] = pyspark_schema
     else:
         options["inferSchema"] = "true"